refactor: Move the test framework code to a framework module
